Extract ClientContextBase
